+++ src/fft_config.rs
@@ -1,7 +1,8 @@
-//! FFT Configuration Macro for Dynamic Setup
+//! FFT Configuration Module for Dynamic Setup
 //!
-//! This module provides a macro to set up FFT-related constants dynamically
-//! based on the microfft library requirements. The macro ensures that:
+//! This module provides macros and utilities for setting up FFT-related constants
+//! and configurations dynamically based on the microfft library requirements.
+//! The module ensures that:
 //! - FFT sizes are powers of 2 (required by microfft's Radix-2 algorithm)
 //! - FFT sizes are within supported range (4 to 32768)
 //! - Related constants are calculated consistently
@@ -28,18 +29,30 @@
 /// - `BIN_WIDTH`: Frequency resolution per bin
 ///
 /// # Example
-/// ```rust
+/// ```rust,no_run
 /// use synthphone_vocals::fft_config;
 ///
-/// // Basic usage with default parameters
-/// fft_config!(1024, 48000.0);
+/// // Basic usage with default parameters (in one scope)
+/// mod basic_config {
+///     use synthphone_vocals::fft_config;
+///     fft_config!(1024, 48000.0);
 ///
-/// // With custom hop ratio and buffer multiplier
-/// fft_config!(2048, 44100.0, hop_ratio = 0.125, buffer_multiplier = 8);
+///     pub fn print_config() {
+///         println!("FFT Size: {}", FFT_SIZE);
+///         println!("Buffer Size: {}", BUFFER_SIZE);
+///     }
+/// }
 ///
-/// // Now you can use SAMPLE_RATE, FFT_SIZE, etc.
-/// println!("FFT Size: {}", FFT_SIZE);
-/// println!("Buffer Size: {}", BUFFER_SIZE);
+/// // With custom parameters (in another scope)
+/// mod custom_config {
+///     use synthphone_vocals::fft_config;
+///     fft_config!(2048, 44100.0, hop_ratio = 0.125, buffer_multiplier = 8);
+///
+///     pub fn print_config() {
+///         println!("FFT Size: {}", FFT_SIZE);
+///         println!("Buffer Size: {}", BUFFER_SIZE);
+///     }
+/// }
 /// ```
 ///
 /// # Microfft Features
@@ -86,7 +99,7 @@ macro_rules! fft_config {
     ($fft_size:expr, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr, block_size = $block_size:expr) => {
         // Compile-time validation that FFT_SIZE is a power of 2
         const _: () = {
-            if !$crate::fft_config::fft_config::is_power_of_two($fft_size) {
+            if !$crate::fft_config::is_power_of_two($fft_size) {
                 panic!("FFT_SIZE must be a power of 2");
             }
             if $fft_size < 4 || $fft_size > 32768 {
@@ -109,7 +122,7 @@ macro_rules! fft_config {
 
         // Compile-time suggestion for optimal microfft features
         const _FFT_FEATURE_SUGGESTION: &str =
-            $crate::fft_config::fft_config::suggest_microfft_feature($fft_size);
+            $crate::fft_config::suggest_microfft_feature($fft_size);
 
         // Optional: Print feature suggestion at compile time (commented out to avoid spam)
         // const _: () = {
@@ -128,74 +141,6 @@ macro_rules! fft_config {
             block_size = 2
         );
     };
-}
-
-/// Helper functions for compile-time validation and suggestions
-pub mod fft_config {
-    /// Check if a number is a power of 2 at compile time
-    pub const fn is_power_of_two(n: usize) -> bool {
-        n > 0 && (n & (n - 1)) == 0
-    }
-
-    /// Suggest the optimal microfft feature for a given FFT size
-    pub const fn suggest_microfft_feature(fft_size: usize) -> &'static str {
-        match fft_size {
-            1..=4 => "size-4",
-            5..=8 => "size-8",
-            9..=16 => "size-16",
-            17..=32 => "size-32",
-            33..=64 => "size-64",
-            65..=128 => "size-128",
-            129..=256 => "size-256",
-            257..=512 => "size-512",
-            513..=1024 => "size-1024",
-            1025..=2048 => "size-2048",
-            2049..=4096 => "size-4096", // This is the default
-            4097..=8192 => "size-8192",
-            8193..=16384 => "size-16384",
-            16385..=32768 => "size-32768",
-            _ => "size-32768", // Maximum supported
-        }
-    }
-
-    /// Get the next power of 2 greater than or equal to n
-    pub const fn next_power_of_two(mut n: usize) -> usize {
-        if n <= 1 {
-            return 1;
-        }
-        n -= 1;
-        n |= n >> 1;
-        n |= n >> 2;
-        n |= n >> 4;
-        n |= n >> 8;
-        n |= n >> 16;
-        #[cfg(target_pointer_width = "64")]
-        {
-            n |= n >> 32;
-        }
-        n + 1
-    }
-
-    /// Validate FFT configuration at runtime
-    pub fn validate_config(
-        fft_size: usize,
-        sample_rate: f32,
-        hop_ratio: f32,
-    ) -> Result<(), &'static str> {
-        if !is_power_of_two(fft_size) {
-            return Err("FFT size must be a power of 2");
-        }
-        if fft_size < 4 || fft_size > 32768 {
-            return Err("FFT size must be between 4 and 32768");
-        }
-        if sample_rate <= 0.0 {
-            return Err("Sample rate must be positive");
-        }
-        if hop_ratio <= 0.0 || hop_ratio > 1.0 {
-            return Err("Hop ratio must be between 0.0 and 1.0");
-        }
-        Ok(())
-    }
 }
 
 /// Convenience macro to create a configuration struct instead of constants
@@ -218,7 +163,7 @@ macro_rules! fft_config_struct {
             // Compile-time validation
             const _VALIDATE: () = {
                 assert!(
-                    $crate::fft_config::fft_config::is_power_of_two($fft_size),
+                    $crate::fft_config::is_power_of_two($fft_size),
                     "FFT_SIZE must be a power of 2"
                 );
                 assert!(
@@ -245,7 +190,7 @@ macro_rules! fft_config_struct {
             // Compile-time validation
             const _VALIDATE: () = {
                 assert!(
-                    $crate::fft_config::fft_config::is_power_of_two($fft_size),
+                    $crate::fft_config::is_power_of_two($fft_size),
                     "FFT_SIZE must be a power of 2"
                 );
                 assert!(
@@ -257,9 +202,144 @@ macro_rules! fft_config_struct {
     };
 }
 
+/// Check if a number is a power of 2 at compile time
+pub const fn is_power_of_two(n: usize) -> bool {
+    n > 0 && (n & (n - 1)) == 0
+}
+
+/// Suggest the optimal microfft feature for a given FFT size
+pub const fn suggest_microfft_feature(fft_size: usize) -> &'static str {
+    match fft_size {
+        1..=4 => "size-4",
+        5..=8 => "size-8",
+        9..=16 => "size-16",
+        17..=32 => "size-32",
+        33..=64 => "size-64",
+        65..=128 => "size-128",
+        129..=256 => "size-256",
+        257..=512 => "size-512",
+        513..=1024 => "size-1024",
+        1025..=2048 => "size-2048",
+        2049..=4096 => "size-4096", // This is the default
+        4097..=8192 => "size-8192",
+        8193..=16384 => "size-16384",
+        16385..=32768 => "size-32768",
+        _ => "size-32768", // Maximum supported
+    }
+}
+
+/// Get the next power of 2 greater than or equal to n
+pub const fn next_power_of_two(mut n: usize) -> usize {
+    if n <= 1 {
+        return 1;
+    }
+    n -= 1;
+    n |= n >> 1;
+    n |= n >> 2;
+    n |= n >> 4;
+    n |= n >> 8;
+    n |= n >> 16;
+    #[cfg(target_pointer_width = "64")]
+    {
+        n |= n >> 32;
+    }
+    n + 1
+}
+
+/// Validate FFT configuration at runtime
+pub fn validate_config(
+    fft_size: usize,
+    sample_rate: f32,
+    hop_ratio: f32,
+) -> Result<(), &'static str> {
+    if !is_power_of_two(fft_size) {
+        return Err("FFT size must be a power of 2");
+    }
+    if fft_size < 4 || fft_size > 32768 {
+        return Err("FFT size must be between 4 and 32768");
+    }
+    if sample_rate <= 0.0 {
+        return Err("Sample rate must be positive");
+    }
+    if hop_ratio <= 0.0 || hop_ratio > 1.0 {
+        return Err("Hop ratio must be between 0.0 and 1.0");
+    }
+    Ok(())
+}
+
+/// Configuration struct for FFT parameters
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub struct FFTConfig {
+    pub fft_size: usize,
+    pub sample_rate: f32,
+    pub hop_ratio: f32,
+    pub buffer_multiplier: usize,
+    pub block_size: usize,
+}
+
+impl FFTConfig {
+    /// Create a new FFT configuration with validation
+    pub fn new(
+        fft_size: usize,
+        sample_rate: f32,
+        hop_ratio: f32,
+        buffer_multiplier: usize,
+        block_size: usize,
+    ) -> Result<Self, &'static str> {
+        validate_config(fft_size, sample_rate, hop_ratio)?;
+
+        if buffer_multiplier < 1 {
+            return Err("buffer_multiplier must be at least 1");
+        }
+
+        if block_size < 1 {
+            return Err("block_size must be at least 1");
+        }
+
+        Ok(Self { fft_size, sample_rate, hop_ratio, buffer_multiplier, block_size })
+    }
+
+    /// Create a default configuration for the given FFT size and sample rate
+    pub fn default_for(fft_size: usize, sample_rate: f32) -> Result<Self, &'static str> {
+        Self::new(fft_size, sample_rate, 0.25, 4, 2)
+    }
+
+    /// Get the buffer size for this configuration
+    pub const fn buffer_size(&self) -> usize {
+        self.fft_size * self.buffer_multiplier
+    }
+
+    /// Get the hop size for this configuration
+    pub fn hop_size(&self) -> usize {
+        (self.fft_size as f32 * self.hop_ratio) as usize
+    }
+
+    /// Get the frequency bin width for this configuration
+    pub fn bin_width(&self) -> f32 {
+        self.sample_rate / self.fft_size as f32
+    }
+
+    /// Get the suggested microfft feature for this configuration
+    pub const fn suggested_microfft_feature(&self) -> &'static str {
+        suggest_microfft_feature(self.fft_size)
+    }
+}
+
+impl Default for FFTConfig {
+    fn default() -> Self {
+        Self {
+            fft_size: 1024,
+            sample_rate: 48000.0,
+            hop_ratio: 0.25,
+            buffer_multiplier: 4,
+            block_size: 2,
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use super::fft_config::*;
+    use super::*;
 
     #[test]
     fn test_is_power_of_two() {
@@ -308,6 +388,30 @@ mod tests {
     }
 
     #[test]
+    fn test_fft_config_struct() {
+        let config = FFTConfig::new(1024, 48000.0, 0.25, 4, 2).unwrap();
+
+        assert_eq!(config.fft_size, 1024);
+        assert_eq!(config.sample_rate, 48000.0);
+        assert_eq!(config.hop_ratio, 0.25);
+        assert_eq!(config.buffer_size(), 4096);
+        assert_eq!(config.hop_size(), 256);
+        assert_eq!(config.bin_width(), 46.875);
+        assert_eq!(config.suggested_microfft_feature(), "size-1024");
+    }
+
+    #[test]
+    fn test_fft_config_default_for() {
+        let config = FFTConfig::default_for(2048, 44100.0).unwrap();
+
+        assert_eq!(config.fft_size, 2048);
+        assert_eq!(config.sample_rate, 44100.0);
+        assert_eq!(config.hop_ratio, 0.25);
+        assert_eq!(config.buffer_multiplier, 4);
+        assert_eq!(config.block_size, 2);
+    }
+
+    #[test]
     fn test_basic_fft_config_macro() {
         // Test basic macro usage
         mod test_basic {
@@ -341,390 +445,4 @@ mod tests {
         assert_eq!(TestConfigWithHop::FFT_SIZE, 1024);
         assert_eq!(TestConfigWithHop::HOP_SIZE, 128); // 1024 * 0.125
     }
-
-    #[test]
-    fn test_process_vocal_effects_config_macros() {
-        // Note: These functions are generated but only for testing compilation
-        // They cannot be called directly in tests due to scope limitations
-        // Test that macro compilation succeeds
-        // Note: Actual function calls are not possible in this test context
-        // due to macro expansion happening at module level
-        assert!(true, "Macro compilation test passed");
-    }
-
-    #[test]
-    fn test_process_vocal_effects_configs_convenience_macro() {
-        // Test that convenience macro compilation succeeds
-        assert!(true, "Convenience macro compilation test passed");
-    }
-
-    #[test]
-    fn test_process_vocal_effects_with_different_settings() {
-        // Test that settings-based functions compile correctly
-        assert!(true, "Settings test compilation passed");
-    }
 }
-
-/// Macro to generate configurable process_vocal_effects_audio functions
-///
-/// This macro creates a version of the `process_vocal_effects_audio` function with custom FFT configuration.
-/// It generates a complete autotune implementation using the specified parameters.
-///
-/// # Arguments
-/// - `$func_name`: Name of the generated function
-/// - `$fft_size`: FFT size (must be power of 2, between 512-4096)
-/// - `$sample_rate`: Sample rate in Hz
-/// - `hop_ratio`: Optional hop size as fraction of FFT size (default: 0.25)
-/// - `buffer_multiplier`: Optional buffer size multiplier (default: 4)
-/// - `block_size`: Optional block size (default: 2)
-///
-/// # Generated Function Signature
-/// ```rust
-/// pub fn $func_name(
-///     unwrapped_buffer: &mut [f32; FFT_SIZE],
-///     last_input_phases: &mut [f32; FFT_SIZE],
-///     last_output_phases: &mut [f32; FFT_SIZE],
-///     previous_pitch_shift_ratio: f32,
-///     config: &AutotuneConfig,
-///     settings: &MusicalSettings,
-/// ) -> [f32; FFT_SIZE]
-/// ```
-///
-/// # Example
-/// ```rust
-/// use synthphone_vocals::{process_vocal_effects_config, AutotuneConfig, MusicalSettings};
-///
-/// // Generate real-time autotune (low latency)
-/// process_vocal_effects_config!(process_vocal_effects_realtime, 512, 48000.0, hop_ratio = 0.5);
-///
-/// // Generate high-quality autotune
-/// process_vocal_effects_config!(process_vocal_effects_hifi, 2048, 48000.0, hop_ratio = 0.125);
-///
-/// let mut buffer = [0.0f32; 512];
-/// let mut input_phases = [0.0f32; 512];
-/// let mut output_phases = [0.0f32; 512];
-/// let config = AutotuneConfig::default();
-/// let settings = MusicalSettings::default();
-///
-/// let result = process_vocal_effects_realtime(
-///     &mut buffer,
-///     &mut input_phases,
-///     &mut output_phases,
-///     1.0,
-///     &config,
-///     &settings
-/// );
-/// ```
-#[macro_export]
-macro_rules! process_vocal_effects_config {
-    // Basic version with just FFT size and sample rate - 512
-    ($func_name:ident, 512, $sample_rate:expr) => {
-        $crate::process_vocal_effects_config_impl_512!($func_name, $sample_rate, 0.25);
-    };
-
-    // Basic version with just FFT size and sample rate - 1024
-    ($func_name:ident, 1024, $sample_rate:expr) => {
-        $crate::process_vocal_effects_config_impl_1024!($func_name, $sample_rate, 0.25);
-    };
-
-    // Basic version with just FFT size and sample rate - 2048
-    ($func_name:ident, 2048, $sample_rate:expr) => {
-        $crate::process_vocal_effects_config_impl_2048!($func_name, $sample_rate, 0.25);
-    };
-
-    // Basic version with just FFT size and sample rate - 4096
-    ($func_name:ident, 4096, $sample_rate:expr) => {
-        $crate::process_vocal_effects_config_impl_4096!($func_name, $sample_rate, 0.25);
-    };
-
-    // Version with hop ratio - 512
-    ($func_name:ident, 512, $sample_rate:expr, hop_ratio = $hop_ratio:expr) => {
-        $crate::process_vocal_effects_config_impl_512!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with hop ratio - 1024
-    ($func_name:ident, 1024, $sample_rate:expr, hop_ratio = $hop_ratio:expr) => {
-        $crate::process_vocal_effects_config_impl_1024!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with hop ratio - 2048
-    ($func_name:ident, 2048, $sample_rate:expr, hop_ratio = $hop_ratio:expr) => {
-        $crate::process_vocal_effects_config_impl_2048!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with hop ratio - 4096
-    ($func_name:ident, 4096, $sample_rate:expr, hop_ratio = $hop_ratio:expr) => {
-        $crate::process_vocal_effects_config_impl_4096!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with buffer multiplier (ignored) - 512
-    ($func_name:ident, 512, $sample_rate:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_512!($func_name, $sample_rate, 0.25);
-    };
-
-    // Version with buffer multiplier (ignored) - 1024
-    ($func_name:ident, 1024, $sample_rate:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_1024!($func_name, $sample_rate, 0.25);
-    };
-
-    // Version with buffer multiplier (ignored) - 2048
-    ($func_name:ident, 2048, $sample_rate:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_2048!($func_name, $sample_rate, 0.25);
-    };
-
-    // Version with buffer multiplier (ignored) - 4096
-    ($func_name:ident, 4096, $sample_rate:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_4096!($func_name, $sample_rate, 0.25);
-    };
-
-    // Full version with all parameters - 512-point FFT
-    ($func_name:ident, 512, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr, block_size = $block_size:expr) => {
-        $crate::process_vocal_effects_config_impl_512!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Full version with all parameters - 1024-point FFT
-    ($func_name:ident, 1024, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr, block_size = $block_size:expr) => {
-        $crate::process_vocal_effects_config_impl_1024!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Full version with all parameters - 2048-point FFT
-    ($func_name:ident, 2048, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr, block_size = $block_size:expr) => {
-        $crate::process_vocal_effects_config_impl_2048!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Full version with all parameters - 4096-point FFT
-    ($func_name:ident, 4096, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr, block_size = $block_size:expr) => {
-        $crate::process_vocal_effects_config_impl_4096!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with named parameters - 512
-    ($func_name:ident, 512, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_512!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with named parameters - 1024
-    ($func_name:ident, 1024, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_1024!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with named parameters - 2048
-    ($func_name:ident, 2048, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_2048!($func_name, $sample_rate, $hop_ratio);
-    };
-
-    // Version with named parameters - 4096
-    ($func_name:ident, 4096, $sample_rate:expr, hop_ratio = $hop_ratio:expr, buffer_multiplier = $buffer_multiplier:expr) => {
-        $crate::process_vocal_effects_config_impl_4096!($func_name, $sample_rate, $hop_ratio);
-    };
-}
-
-/// Implementation macro for 512-point FFT autotune function
-#[macro_export]
-macro_rules! process_vocal_effects_config_impl_512 {
-    ($func_name:ident, $sample_rate:expr, $hop_ratio:expr) => {
-        pub fn $func_name(
-            unwrapped_buffer: &mut [f32; 512],
-            last_input_phases: &mut [f32; 512],
-            last_output_phases: &mut [f32; 512],
-            previous_pitch_shift_ratio: f32,
-            config: &$crate::AutotuneConfig,
-            settings: &$crate::MusicalSettings,
-        ) -> [f32; 512] {
-            $crate::process_vocal_effects_core::process_vocal_effects_512(
-                unwrapped_buffer,
-                last_input_phases,
-                last_output_phases,
-                previous_pitch_shift_ratio,
-                config,
-                settings,
-                $sample_rate,
-                $hop_ratio,
-            )
-        }
-    };
-}
-
-/// Implementation macro for 1024-point FFT autotune function
-#[macro_export]
-macro_rules! process_vocal_effects_config_impl_1024 {
-    ($func_name:ident, $sample_rate:expr, $hop_ratio:expr) => {
-        pub fn $func_name(
-            unwrapped_buffer: &mut [f32; 1024],
-            last_input_phases: &mut [f32; 1024],
-            last_output_phases: &mut [f32; 1024],
-            previous_pitch_shift_ratio: f32,
-            config: &$crate::AutotuneConfig,
-            settings: &$crate::MusicalSettings,
-        ) -> [f32; 1024] {
-            $crate::process_vocal_effects_core::process_vocal_effects_1024(
-                unwrapped_buffer,
-                last_input_phases,
-                last_output_phases,
-                previous_pitch_shift_ratio,
-                config,
-                settings,
-                $sample_rate,
-                $hop_ratio,
-            )
-        }
-    };
-}
-
-/// Implementation macro for 2048-point FFT autotune function
-#[macro_export]
-macro_rules! process_vocal_effects_config_impl_2048 {
-    ($func_name:ident, $sample_rate:expr, $hop_ratio:expr) => {
-        pub fn $func_name(
-            unwrapped_buffer: &mut [f32; 2048],
-            last_input_phases: &mut [f32; 2048],
-            last_output_phases: &mut [f32; 2048],
-            previous_pitch_shift_ratio: f32,
-            config: &$crate::AutotuneConfig,
-            settings: &$crate::MusicalSettings,
-        ) -> [f32; 2048] {
-            $crate::process_vocal_effects_core::process_vocal_effects_2048(
-                unwrapped_buffer,
-                last_input_phases,
-                last_output_phases,
-                previous_pitch_shift_ratio,
-                config,
-                settings,
-                $sample_rate,
-                $hop_ratio,
-            )
-        }
-    };
-}
-
-/// Implementation macro for 4096-point FFT autotune function
-#[macro_export]
-macro_rules! process_vocal_effects_config_impl_4096 {
-    ($func_name:ident, $sample_rate:expr, $hop_ratio:expr) => {
-        pub fn $func_name(
-            unwrapped_buffer: &mut [f32; 4096],
-            last_input_phases: &mut [f32; 4096],
-            last_output_phases: &mut [f32; 4096],
-            previous_pitch_shift_ratio: f32,
-            config: &$crate::AutotuneConfig,
-            settings: &$crate::MusicalSettings,
-        ) -> [f32; 4096] {
-            $crate::process_vocal_effects_core::process_vocal_effects_4096(
-                unwrapped_buffer,
-                last_input_phases,
-                last_output_phases,
-                previous_pitch_shift_ratio,
-                config,
-                settings,
-                $sample_rate,
-                $hop_ratio,
-            )
-        }
-    };
-}
-
-/// Example usage showing how to use the process_vocal_effects_config macro
-///
-/// ```rust,no_run
-/// use synthphone_vocals::{process_vocal_effects_config, AutotuneConfig, MusicalSettings};
-///
-/// // Generate different quality configurations
-/// process_vocal_effects_config!(process_vocal_effects_realtime, 512, 48000.0, hop_ratio = 0.5);
-/// process_vocal_effects_config!(process_vocal_effects_balanced, 1024, 48000.0, hop_ratio = 0.25);
-/// process_vocal_effects_config!(process_vocal_effects_quality, 2048, 48000.0, hop_ratio = 0.125);
-/// process_vocal_effects_config!(process_vocal_effects_hifi, 4096, 48000.0, hop_ratio = 0.0625);
-///
-/// // Usage in your application
-/// fn process_audio() {
-///     let mut buffer = [0.0f32; 1024];
-///     let mut input_phases = [0.0f32; 1024];
-///     let mut output_phases = [0.0f32; 1024];
-///
-///     let config = AutotuneConfig::default();
-///     let mut settings = MusicalSettings::default();
-///     settings.key = 0;  // C major
-///     settings.note = 0; // Auto mode
-///     settings.formant = 1; // Enable formant preservation
-///
-///     // Process with balanced quality
-///     let result = process_vocal_effects_balanced(
-///         &mut buffer,
-///         &mut input_phases,
-///         &mut output_phases,
-///         1.0, // Previous pitch shift ratio
-///         &config,
-///         &settings
-///     );
-/// }
-/// ```
-///
-/// ## Configuration Guidelines
-///
-/// ### FFT Size Selection
-/// - **512**: Ultra-low latency (10.7ms @ 48kHz), minimal CPU, lower quality
-/// - **1024**: Low latency (21.3ms @ 48kHz), good for real-time applications
-/// - **2048**: Balanced (42.7ms @ 48kHz), good quality/performance trade-off
-/// - **4096**: High quality (85.3ms @ 48kHz), best for offline processing
-///
-/// ### Hop Ratio Selection
-/// - **0.0625** (1/16): Highest quality, 93.75% overlap, most CPU intensive
-/// - **0.125** (1/8): Very high quality, 87.5% overlap, high CPU usage
-/// - **0.25** (1/4): Good quality, 75% overlap, moderate CPU usage (default)
-/// - **0.5** (1/2): Lower quality, 50% overlap, lowest CPU usage
-///
-/// ### Formant Settings in MusicalSettings
-/// - **0**: No formant processing (fastest)
-/// - **1**: Lower formants (deeper voice effect)
-/// - **2**: Raise formants (higher voice effect)
-///
-/// ### Performance vs Quality Trade-offs
-///
-/// | Configuration | Latency | CPU Usage | Quality | Best For |
-/// |---------------|---------|-----------|---------|----------|
-/// | 512, 0.5 | Ultra-low | Very Low | Basic | Live performance |
-/// | 1024, 0.25 | Low | Low | Good | Real-time apps |
-/// | 2048, 0.125 | Medium | Medium | High | Studio recording |
-/// | 4096, 0.0625 | High | High | Excellent | Post-processing |
-
-/// Convenience macro to create multiple autotune configurations at once
-///
-/// This macro generates multiple autotune functions with different configurations
-/// in a single call, useful for applications that need different quality levels.
-///
-/// # Example
-/// ```rust
-/// use synthphone_vocals::process_vocal_effects_configs;
-///
-/// process_vocal_effects_configs! {
-///     fast => (process_vocal_effects_fast, 512, 48000.0, hop_ratio = 0.5),
-///     balanced => (process_vocal_effects_balanced, 1024, 48000.0, hop_ratio = 0.25),
-///     quality => (process_vocal_effects_quality, 2048, 48000.0, hop_ratio = 0.125),
-///     hifi => (process_vocal_effects_hifi, 4096, 48000.0, hop_ratio = 0.0625)
-/// }
-///
-/// // Now you can use process_vocal_effects_fast, process_vocal_effects_balanced, etc.
-/// ```
-#[macro_export]
-macro_rules! process_vocal_effects_configs {
-    // Handle mixed FFT sizes by processing each entry individually
-    (@single $name:ident => ($func_name:ident, 512, $sample_rate:expr $(, $param:ident = $value:expr)*)) => {
-        $crate::process_vocal_effects_config!($func_name, 512, $sample_rate $(, $param = $value)*);
-    };
-    (@single $name:ident => ($func_name:ident, 1024, $sample_rate:expr $(, $param:ident = $value:expr)*)) => {
-        $crate::process_vocal_effects_config!($func_name, 1024, $sample_rate $(, $param = $value)*);
-    };
-    (@single $name:ident => ($func_name:ident, 2048, $sample_rate:expr $(, $param:ident = $value:expr)*)) => {
-        $crate::process_vocal_effects_config!($func_name, 2048, $sample_rate $(, $param = $value)*);
-    };
-    (@single $name:ident => ($func_name:ident, 4096, $sample_rate:expr $(, $param:ident = $value:expr)*)) => {
-        $crate::process_vocal_effects_config!($func_name, 4096, $sample_rate $(, $param = $value)*);
-    };
-
-    // Main entry point - delegates to @single
-    ($($name:ident => ($func_name:ident, $fft_size:tt, $sample_rate:expr $(, $param:ident = $value:expr)*)),* $(,)?) => {
-        $(
-            $crate::process_vocal_effects_configs!(@single $name => ($func_name, $fft_size, $sample_rate $(, $param = $value)*));
-        )*
-    };
-}
-
-// Example usage documentation - moved to module level to avoid orphaned doc comment
