+++ src/process_vocal_effects.rs
